Basic website outline
Created the three pages, created both the layout and style sheets for the pages. The page title is center aligned and enlarged. The nav bar is in the center, and balanced on it's sides. The left most aside is where side notes and random facts will go. The right aside is where pictures will most likely go. The article is in the center and needed no css to be where it is. Each page is the same in composition for now but can be edited later. Naturally links have been added, and the comments saying author and date were added to the css sheets but not the html pages. The nav bar links change color and have no text decoration.