More polish, picture, content, etc.
Added a picture to each of the columns, along with a fact for the left side. The article now is divided up and is offset from the sides of it's area. A font has been added to the fun fact side. Content of the main page and the history pages have been added. now need to add more images, web pages, and fulfill more requirements of the project.